requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
drf-yasg>=1.21,<1.22
